refactor: 패키지명 수정 및 compare 로직 변경
